# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ResponseMessage": {
				"properties": {
					"message": {
						"type": "string"
					},
					"detalles": {}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResponseMessageWithData_any_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"detalles": {},
					"data": {}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.EstadoProducto": {
				"type": "string",
				"enum": [
					"DISPONIBLE",
					"AGOTADO",
					"RESERVADO",
					"VENCIDO",
					"ELIMINADO"
				]
			},
			"EstadoProducto": {
				"$ref": "#/components/schemas/_36_Enums.EstadoProducto"
			},
			"ProductosDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"codigoBarras": {
						"type": "string",
						"nullable": true
					},
					"codigoQR": {
						"type": "string",
						"nullable": true
					},
					"nombre": {
						"type": "string"
					},
					"descripcion": {
						"type": "string"
					},
					"cantidad": {
						"type": "number",
						"format": "double"
					},
					"precio": {
						"type": "number",
						"format": "double"
					},
					"fechaAdquisicion": {
						"type": "string"
					},
					"fechaVencimiento": {
						"type": "string"
					},
					"usuarioId": {
						"type": "number",
						"format": "double"
					},
					"estado": {
						"$ref": "#/components/schemas/EstadoProducto"
					},
					"imagen": {
						"type": "string"
					},
					"categoria": {
						"type": "string"
					}
				},
				"required": [
					"nombre",
					"descripcion",
					"cantidad",
					"precio",
					"fechaAdquisicion",
					"fechaVencimiento",
					"usuarioId",
					"estado",
					"imagen"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"legacyJwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "producttrack",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@mariaximena0724 y emilia"
		}
	},
	"paths": {
...
		"/productos": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Obtiene todos los productos con filtros opcionales.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/productos.controller)",
				"tags": [
					"productos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "nombre",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "categoria",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "estado",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fechaAdquisicionDesde",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fechaAdquisicionHasta",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fechaVencimientoDesde",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fechaVencimientoHasta",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Producto creado correctamente",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ResponseMessageWithData_any_"
										},
										{
											"$ref": "#/components/schemas/ResponseMessage"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Datos inválidos"
					}
				},
				"description": "Crea un nuevo producto en la base de datos.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/productos.controller)",
				"tags": [
					"productos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Objeto con los datos del producto",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductosDTO",
								"description": "Objeto con los datos del producto"
							}
						}
					}
				}
			}
		},
		"/productos/categorias": {
			"get": {
				"operationId": "ObtenerCategorias",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										{
											"$ref": "#/components/schemas/ResponseMessage"
										}
									]
								}
							}
						}
					}
				},
				"description": "Obtiene las categorías únicas para los productos según tipoUsuario.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/productos.controller)",
				"tags": [
					"productos"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tipoUsuario",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/productos/por-categoria": {
			"get": {
				"operationId": "GetByCategoria",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Devuelve todos los productos en una categoría específica.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/productos.controller)",
				"tags": [
					"productos"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "categoria",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/productos/nombres/{idUsuario}": {
			"get": {
				"operationId": "GetNombresProductosDelUsuario",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Devuelve los nombres de los productos registrados por un usuario.\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/productos.controller)",
				"tags": [
					"productos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idUsuario",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/productos/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Obtiene un producto por su ID.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/productos.controller)",
				"tags": [
					"productos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProducto",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseMessage"
								}
							}
						}
					}
				},
				"description": "Actualiza un producto existente.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/productos.controller)",
				"tags": [
					"productos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProducto",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseMessage"
								}
							}
						}
					}
				},
				"description": "Elimina un producto de la base de datos.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/productos.controller)",
				"tags": [
					"productos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}